+++ tests/negative-prefix.test.js
@@ -343,4 +343,24 @@ crosscheck(() => {
       return expect(result.css).toMatchCss(css``)
     })
   })
+
+  // This is a weird test but it used to crash because the negative prefix + variant used to cause an undefined utility to be generated
+  test('addUtilities without negative prefix + variant + negative prefix in content should not crash', async () => {
+    let config = {
+      content: [{ raw: html`<div class="hover:-top-lg"></div>` }],
+      plugins: [
+        ({ addUtilities }) => {
+          addUtilities({
+            '.top-lg': {
+              top: '6rem',
+            },
+          })
+        },
+      ],
+    }
+
+    let result = await run('@tailwind utilities', config)
+
+    expect(result.css).toMatchCss(css``)
+  })
 })